changed application to get the startup time from the java app

+++ apps/NeighborDetection/java-src/it/unitn/wildmac/demo/Demo.java
@@ -16,7 +16,7 @@ import it.unitn.wildmac.WSNGateway;
  */
 public class Demo {
 
-	protected static final long EXPERIMENT_DELAY = 2000;
+	protected static final int EXPERIMENT_DELAY = 2000;
 
 	private static void printSyntax() {
 		System.out.println("Syntax:");
